+++ purchases_ui_flutter/ios/Classes/Helpers/UIView+.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+extension UIView {
+    var parentViewController: UIViewController? {
+        var responder: UIResponder? = self
+        while let nextResponder = responder?.next {
+            if let viewController = nextResponder as? UIViewController {
+                return viewController
+            }
+            responder = nextResponder
+        }
+        return nil
+    }
+}
