+++ purchases_ui_flutter/ios/Classes/PurchasesUiPaywallView.swift
@@ -125,19 +125,6 @@ class PurchasesUiPaywallView: NSObject, FlutterPlatformView {
     }
 }
 
-extension UIView {
-    var parentViewController: UIViewController? {
-        var responder: UIResponder? = self
-        while let nextResponder = responder?.next {
-            if let viewController = nextResponder as? UIViewController {
-                return viewController
-            }
-            responder = nextResponder
-        }
-        return nil
-    }
-}
-
 @available(iOS 15.0, *)
 extension PurchasesUiPaywallView: PaywallViewControllerDelegateWrapper {
     func paywallViewController(_ controller: PaywallViewController,
